requirements.txt
paho-mqtt==1.5.0
PyYAML==5.3.1
revpimodio2==2.5.3
